Map hyper+tab to ctrl+tab
This got a bit crazy. I'm not entirely sure what's going on, but sending
ctrl+tab from a callback function doesn't seem to work when mapping the
hyper key to `["cmd" "ctrl" "alt" "shift"]`. So what we do here is map
the hyper key to `"f19"` trough a modal hotkey with no actual trigger.
We bind the pressed/released fns to the enter/exit functions of the
modal.

This works pretty well because we no longer need Karabiner. This can be
done with standard Mac Settings. This file will map capslock to f19

~/Library/LaunchAgents/com.local.KeyRemapping.plist
```plist
<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
<plist version="1.0">
<dict>
    <key>Label</key>
    <string>com.local.KeyRemapping</string>
    <key>ProgramArguments</key>
    <array>
        <string>/usr/bin/hidutil</string>
        <string>property</string>
        <string>--set</string>
        <string>{"UserKeyMapping":[
            {
              "HIDKeyboardModifierMappingSrc": 0x700000039,
              "HIDKeyboardModifierMappingDst": 0x70000006E
            }
        ]}</string>
    </array>
    <key>RunAtLoad</key>
    <true/>
</dict>
</plist>
```

Also snuck in some niceties around dark mode. Somehow I got stuck in
dark mode and the font contrast was terrible. So I just set it to light
mode all the time. Let's fix this right away.